pyproject.toml
[tool.poetry]
name = "automated-flood-extent-mapping"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{ include = "flood_mapping", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
rasterio = "^1.3.9"
matplotlib = "^3.8.1"
sentinelhub = "^3.9.2"
numpy = "^1.26.1"
pandas = "^2.1.2"
geopandas = "^0.14.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
pre-commit = "^3.5.0"
ruff = "^0.1.5"       # should be same as in .pre-commit-config.yaml

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 110
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N"]
ignore = ["E402", "E501"]
